Se realiza refactor en el controller de facturas. Se mapean las excepciones correctamente al momento de responder un request (aun no se implemento la generacion de un pago en dicho controller).

+++ src/main/java/com/realstate/services/RentalBillService.java
@@ -65,11 +65,17 @@ public class RentalBillService {
 		return rentalBillRepository.insert(rentalBillList);
 	}
 	
-	public RentalBill update(RentalBill rentalBill) {
+	public RentalBill update(RentalBill rentalBill) throws RentalBillDoesNotExistException {
+		if (!rentalBillRepository.existsById(new ObjectId(rentalBill.getRentalBillId()))) {
+			throw new RentalBillDoesNotExistException();
+		}
 		return rentalBillRepository.save(rentalBill);
 	}
 	
-	public void delete(RentalBill rentalBill) {
+	public void delete(RentalBill rentalBill) throws RentalBillDoesNotExistException {
+		if (!rentalBillRepository.existsById(new ObjectId(rentalBill.getRentalBillId()))) {
+			throw new RentalBillDoesNotExistException();
+		}
 		rentalBillRepository.delete(rentalBill);
 	}
